make more fair

+++ benches/roundup.rs
@@ -10,7 +10,7 @@ fn benchmark_encoding(c: &mut Criterion) {
     // Generate random u128 values from random u32 values
     let mut rng = rng();
     let values: Vec<u128> = (0..100)
-        .map(|_| rng.random_range(0..u32::MAX) as u128)
+        .map(|_| rng.random_range(0..u128::MAX) as u128)
         .collect();
 
     // Benchmark Borsh encoding
